検索画面にMVVMを適用
