+++ iOSEngineerCodeCheck/SearchPage/MainModel.swift
@@ -1,0 +1,11 @@
+//
+//  MainModel.swift
+//  iOSEngineerCodeCheck
+//
+//  Created by Shingo Toyoda on 2024/04/06.
+//  Copyright © 2024 YUMEMI Inc. All rights reserved.
+//
+
+struct Repository {
+    let fullName: String
+}
