+++ iOSEngineerCodeCheck/SearchPage/MainViewController.swift
@@ -1,17 +1,11 @@
-//
-//  MainViewController.swift
-//  iOSEngineerCodeCheck
-//
-//  Created by 史 翔新 on 2020/04/20.
-//  Copyright © 2020 YUMEMI Inc. All rights reserved.
-//
-
 import UIKit
+import Combine
 
 class MainViewController: UITableViewController {
-
     @IBOutlet weak var searchBar: UISearchBar!
-    
+
+    private var cancellables: Set<AnyCancellable> = []
+    private var viewModel: MainViewModel = MainViewModel()
     private var dataSource: GitHubReposDataSource = GitHubReposDataSource()
 
     override func viewDidLoad() {
@@ -19,55 +13,36 @@ class MainViewController: UITableViewController {
         searchBar.placeholder = Constant.searchBarPlaceholder
         searchBar.delegate = self
         tableView.dataSource = dataSource
+
+        viewModel.$repos
+            .receive(on: DispatchQueue.main)
+            .sink { [weak self] _ in
+                self?.dataSource.update(with: self?.viewModel.repos ?? [])
+                self?.tableView.reloadData()
+            }
+            .store(in: &cancellables)
     }
 
     override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        if segue.identifier == "Detail" {
-            if let detail = segue.destination as? DetailViewController,
-               // RepoItemはsenderから受け取るほうが安全
-               let repo = sender as? Repository {
-                detail.viewModel.fullName = repo.fullName
-            }
+        if segue.identifier == "Detail",
+           let detail = segue.destination as? DetailViewController,
+           let repo = sender as? Repository {
+            detail.viewModel.fullName = repo.fullName
         }
     }
 
     override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        // セルをタップしたときに呼ばれる
-        if let repo = dataSource.repo(at: indexPath.row) {
+        if let repo = viewModel.repos[safe: indexPath.row] {
             performSegue(withIdentifier: "Detail", sender: repo)
         }
     }
 }
 
 extension MainViewController: UISearchBarDelegate {
-    
-    func searchBarShouldBeginEditing(_ searchBar: UISearchBar) -> Bool {
-        return true
-    }
-    
     func searchBarSearchButtonClicked(_ searchBar: UISearchBar) {
         guard let keyword = searchBar.text, !keyword.isEmpty else {
             return
         }
-        if !keyword.isEmpty {
-            let apiClient: APIClient = APIClient(baseURL: URL(string: Constant.githubAPIURL)!)
-            let request: SearchReposRequest = SearchReposRequest(keyword: keyword)
-            apiClient.send(request) {[weak self] result in
-                guard let self = self else {
-                    return
-                }
-                switch result {
-                case .success(let response):
-                    let repos: [Repository] = response.items.map { Repository(fullName: $0.fullName) }
-                    DispatchQueue.main.async {
-                        self.dataSource.update(with: repos)
-                        self.tableView.reloadData()
-                    }
-                case .failure(let error):
-                    print(error)
-                }
-            }
-        }
+        viewModel.searchRepos(with: keyword)
     }
-
 }
